pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-paradedb"
version = "0.1.0"
description = "ParadeDB full-text search integration for Django ORM"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [
    { name = "ParadeDB", email = "support@paradedb.com" },
]
keywords = [
    "django",
    "paradedb",
    "postgres",
    "postgresql",
    "search",
    "bm25",
    "text search",
    "full text search",
    "full-text search",
    "hybrid search",
    "faceted search",
    "facets",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "Django>=6.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.8",
    "ruff>=0.8",
    "mypy>=1.13",
    "django-stubs>=5.1",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "psycopg[binary]>=3.1",
]
[project.urls]
Homepage = "https://github.com/paradedb/django-paradedb"
Documentation = "https://docs.paradedb.com"
Repository = "https://github.com/paradedb/django-paradedb"
Issues = "https://github.com/paradedb/django-paradedb/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/paradedb"]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "DJ",     # flake8-django
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["paradedb"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "tests.settings"
strict_settings = false

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]
django_find_project = false
DJANGO_SETTINGS_MODULE = "tests.settings"
addopts = [
    "--strict-markers",
    "-ra",
]
markers = [
    "integration: tests that require a ParadeDB Postgres instance",
]

[tool.coverage.run]
source = ["src/paradedb"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
